refactor: extract shared OGC API discovery into base/_ogc_api
The /collections discovery (URL builder, LRU-cached fetch, id/error/HTTP-error
parsing) and the GDAL HTTP config were private to feature/_oapif; the new
Dataset.from_ogc_coverages reader needs the same primitives. Hoist them into
pyramids.base._ogc_api (one copy) and import them into _oapif/_wfs, so both the
Features and Coverages readers share the discovery instead of duplicating it.
