+++ src/pyramids/feature/_ogc.py
@@ -1,29 +1,16 @@
-"""Shared primitives for the OGC vector readers (WFS and OGC API – Features).
+"""Shared read-filter assembly for the OGC vector readers (WFS and OGC API – Features).
 
 `pyramids.feature._wfs` and `pyramids.feature._oapif` are sibling factories that
 both drive a GDAL OGR HTTP driver behind `FeatureCollection`. The read-filter
-assembly and the GDAL HTTP config (auth + timeout) are identical between them, so
-they live here once instead of being copied into each reader.
+assembly (bbox / attribute filter / feature cap) is identical between them, so it
+lives here once instead of being copied into each reader. The GDAL HTTP config
+(auth + timeout) is shared more widely with the raster OGC readers and lives in
+:mod:`pyramids.base._ogc_api`.
 """
 
 from __future__ import annotations
 
 from typing import Any
-
-# GDAL's HTTP Basic-auth env var. Assembled in two pieces so static analysis does
-# not misread the literal key as a hard-coded credential: the value is always
-# supplied by the caller's ``auth``, never hard-coded here.
-GDAL_HTTP_AUTH_VAR = "GDAL_HTTP_USER" + "PWD"
-
-
-def gdal_http_config(auth: tuple[str, str] | None, timeout: float) -> dict[str, str]:
-    """GDAL config options for the OGC HTTP requests (auth + timeout)."""
-    # GDAL_HTTP_TIMEOUT is whole seconds; clamp to >= 1 so a sub-second timeout is
-    # not truncated to "0", which GDAL reads as "no timeout".
-    config = {"GDAL_HTTP_TIMEOUT": str(max(1, int(timeout)))}
-    if auth is not None:
-        config[GDAL_HTTP_AUTH_VAR] = f"{auth[0]}:{auth[1]}"
-    return config
 
 
 def read_kwargs(
